+++ packages/gatsby-transformer-remark/src/__tests__/on-node-create.js
@@ -1,4 +1,3 @@
-const Promise = require(`bluebird`)
 const _ = require(`lodash`)
 
 const onCreateNode = require(`../on-node-create`)
@@ -119,6 +118,50 @@ yadda yadda
       })
 
       expect(parsed.frontmatter.date).toEqual(new Date(date).toJSON())
+    })
+
+    it(`Filters nodes with the given filter function, if provided`, async () => {
+      const content = ``
+
+      node.content = content
+      node.sourceInstanceName = `gatsby-test-source`
+
+      const createNode = jest.fn()
+      const createParentChildLink = jest.fn()
+      const actions = { createNode, createParentChildLink }
+      const createNodeId = jest.fn()
+      createNodeId.mockReturnValue(`uuid-from-gatsby`)
+
+      await onCreateNode(
+        {
+          node,
+          loadNodeContent,
+          actions,
+          createNodeId,
+        },
+        {
+          filter: node =>
+            node.sourceInstanceName === `gatsby-other-test-source`,
+        }
+      ).then(() => {
+        expect(createNode).toHaveBeenCalledTimes(0)
+        expect(createParentChildLink).toHaveBeenCalledTimes(0)
+      })
+
+      await onCreateNode(
+        {
+          node,
+          loadNodeContent,
+          actions,
+          createNodeId,
+        },
+        {
+          filter: node => node.sourceInstanceName === `gatsby-test-source`,
+        }
+      ).then(() => {
+        expect(createNode).toHaveBeenCalledTimes(1)
+        expect(createParentChildLink).toHaveBeenCalledTimes(1)
+      })
     })
   })