+++ www/src/components/function-list.js
@@ -2,6 +2,7 @@ import { graphql } from "gatsby"
 import React from "react"
 
 import { rhythm, scale, options } from "../utils/typography"
+import presets, { space, colors } from "../utils/presets"
 
 const Param = (param, depth = 0) => {
   // The "plugin" parameter is used internally but not
@@ -14,7 +15,7 @@ const Param = (param, depth = 0) => {
     <div
       key={`param ${JSON.stringify(param)}`}
       css={{
-        marginLeft: `${depth * 1.05}rem`,
+        marginLeft: `${rhythm(depth * 1)}`,
         ...(depth > 0 && scale((depth === 1 ? -1 : -1.5) / 5)),
         lineHeight: options.baseLineHeight,
       }}
@@ -28,10 +29,12 @@ const Param = (param, depth = 0) => {
         {param.name === `$0` ? `destructured object` : param.name}
         {` `}
         {param.type && param.name !== `$0` && (
-          <span css={{ color: `#73725f` }}>{`{${param.type.name}}`}</span>
+          <span css={{ color: colors.gray.calm, fontWeight: `normal` }}>{`{${
+            param.type.name
+          }}`}</span>
         )}
         {param.default && (
-          <span css={{ color: `#73725f` }}>
+          <span css={{ color: colors.gray.calm, fontWeight: `normal` }}>
             [default=
             {param.default}]
           </span>
@@ -39,14 +42,18 @@ const Param = (param, depth = 0) => {
       </h5>
       {param.description && (
         <div
-          css={{ marginBottom: rhythm(-1 / 4) }}
           dangerouslySetInnerHTML={{
             __html: param.description.childMarkdownRemark.html,
           }}
         />
       )}
       {param.properties && (
-        <div css={{ marginBottom: rhythm(1), marginTop: rhythm(1 / 2) }}>
+        <div
+          css={{
+            marginBottom: rhythm(space[5]),
+            marginTop: rhythm(space[3]),
+          }}
+        >
           {param.properties.map(param => Param(param, depth + 1))}
         </div>
       )}
@@ -60,7 +67,7 @@ export default ({ functions }) => (
       <div
         id={node.name}
         key={`reference list ${node.name}`}
-        css={{ marginBottom: rhythm(1) }}
+        css={{ marginBottom: rhythm(space[5]) }}
       >
         {i !== 0 && <hr />}
         <h3>
@@ -86,13 +93,13 @@ export default ({ functions }) => (
               <div
                 key={`ret ${JSON.stringify(ret)}`}
                 css={{
-                  marginLeft: `1.05rem`,
-                  ...scale(-1 / 5),
+                  marginLeft: rhythm(space[4]),
+                  fontSize: presets.scale[1],
                   lineHeight: options.baseLineHeight,
                 }}
               >
                 <h5 css={{ margin: 0 }}>
-                  <span css={{ color: `#73725f` }}>
+                  <span css={{ color: colors.gray.calm }}>
                     {`{${
                       ret.type.type === `UnionType`
                         ? ret.type.elements.map(el => String(el.name)).join(`|`)
@@ -102,7 +109,6 @@ export default ({ functions }) => (
                 </h5>
                 {ret.description && (
                   <div
-                    css={{ marginBottom: rhythm(-1 / 4) }}
                     dangerouslySetInnerHTML={{
                       __html: ret.description.childMarkdownRemark.html,
                     }}
@@ -115,7 +121,7 @@ export default ({ functions }) => (
 
         {node.examples && node.examples.length > 0 && (
           <div>
-            <h4 css={{ marginTop: rhythm(1) }}>Example</h4>
+            <h4 css={{ marginTop: rhythm(space[5]) }}>Example</h4>
             {` `}
             {node.examples.map((example, i) => (
               <div