+++ www/src/components/navigation-mobile.js
@@ -8,9 +8,9 @@ import {
   PluginsIcon,
   ShowcaseIcon,
 } from "../assets/mobile-nav-icons"
-import presets, { colors } from "../utils/presets"
+import presets, { colors, transition, radii, space } from "../utils/presets"
 import { svgStyles } from "../utils/styles"
-import typography, { rhythm, scale, options } from "../utils/typography"
+import { rhythm, options } from "../utils/typography"
 
 const getProps = ({ isPartiallyCurrent }) => {
   return {
@@ -41,7 +41,7 @@ const MobileNavItem = ({ linkTo, label, icon }) => (
 )
 
 const MobileNavigation = () => (
-  <React.Fragment>
+  <>
     <span
       css={{
         position: `absolute`,
@@ -65,10 +65,10 @@ const MobileNavigation = () => (
         left: 0,
         right: 0,
         zIndex: 1,
-        borderTop: `1px solid ${colors.ui.border}`,
-        background: colors.ui.whisper,
-        minHeight: presets.headerHeight,
-        fontFamily: typography.options.headerFontFamily.join(`,`),
+        borderTop: `1px solid ${colors.ui.light}`,
+        background: colors.white,
+        height: presets.headerHeight,
+        fontFamily: options.headerFontFamily.join(`,`),
         paddingBottom: `env(safe-area-inset-bottom)`,
         [presets.Md]: {
           display: `none`,
@@ -81,7 +81,7 @@ const MobileNavigation = () => (
       <MobileNavItem linkTo="/blog/" label="Blog" icon={BlogIcon} />
       <MobileNavItem linkTo="/showcase/" label="Showcase" icon={ShowcaseIcon} />
     </div>
-  </React.Fragment>
+  </>
 )
 
 export default MobileNavigation
@@ -101,12 +101,12 @@ const styles = {
       "& .svg-stroke-lilac": { stroke: colors.lavender },
       "& .svg-fill-lilac": { fill: colors.lavender },
       "& .svg-fill-gatsby": { fill: colors.lavender },
-      "& .svg-fill-brightest": { fill: `#fff` },
+      "& .svg-fill-brightest": { fill: colors.white },
       "& .svg-fill-accent": { fill: colors.lavender },
       "& .svg-stroke-gatsby": { stroke: colors.lavender },
       "& .svg-fill-gradient-accent-white-top": { fill: `transparent` },
       "& .svg-fill-gradient-accent-white-45deg": { fill: `transparent` },
-      "& .svg-fill-gradient-accent-white-bottom": { fill: `#fff` },
+      "& .svg-fill-gradient-accent-white-bottom": { fill: colors.white },
       "& .svg-fill-gradient-purple": { fill: colors.lavender },
       "& .svg-stroke-gradient-purple": { stroke: colors.lavender },
       "& .svg-fill-wisteria": { fill: `transparent` },
@@ -116,34 +116,30 @@ const styles = {
   },
   link: {
     default: {
-      color: colors.gatsby,
-      borderRadius: presets.radius,
-      fontSize: scale(-1 / 2).fontSize,
+      color: colors.lilac,
+      borderRadius: radii[1],
+      fontSize: presets.scale[0],
       flexShrink: 0,
-      lineHeight: 1,
+      lineHeight: presets.lineHeights.solid,
       width: 64,
-      padding: `${rhythm(options.blockMarginBottom / 4)} ${rhythm(
-        options.blockMarginBottom / 4
-      )} 0`,
+      padding: rhythm(space[1]),
       textDecoration: `none`,
       textAlign: `center`,
+      WebkitFontSmoothing: `antialiased`,
       "& svg": {
         display: `block`,
         height: 32,
         margin: `0 auto`,
         "& path, & line, & polygon": {
-          transition: `all ${presets.animation.speedDefault} ${
-            presets.animation.curveDefault
+          transition: `all ${transition.speed.default} ${
+            transition.curve.default
           }`,
         },
       },
     },
     active: {
-      "&&": {
-        color: colors.gatsby,
-        fontWeight: `bold`,
-        // WebkitFontSmoothing: `antialiased`,
-      },
+      color: colors.gatsby,
+      fontWeight: `bold`,
     },
   },
 }