+++ www/src/utils/styles.js
@@ -1,8 +1,8 @@
 import hex2rgba from "hex2rgba"
 import { keyframes } from "@emotion/core"
 
-import presets, { colors } from "./presets"
-import { rhythm, scale, options } from "./typography"
+import presets, { colors, space, transition, radii } from "./presets"
+import { rhythm, options } from "./typography"
 
 const stripeAnimation = keyframes({
   "0%": { backgroundPosition: `0 0` },
@@ -12,8 +12,8 @@ const stripeAnimation = keyframes({
 export const scrollbarStyles = {
   WebkitOverflowScrolling: `touch`,
   "&::-webkit-scrollbar": {
-    width: `6px`,
-    height: `6px`,
+    width: rhythm(space[2]),
+    height: rhythm(space[2]),
   },
   "&::-webkit-scrollbar-thumb": {
     background: colors.ui.bright,
@@ -30,43 +30,36 @@ export const buttonStyles = {
   default: {
     alignItems: `center`,
     backgroundColor: colors.gatsby,
-    borderRadius: presets.radius,
+    borderRadius: radii[1],
     borderWidth: 1,
     borderStyle: `solid`,
     borderColor: colors.gatsby,
-    boxShadow: `none`,
-    color: `#fff`,
+    color: colors.white,
     cursor: `pointer`,
     display: `inline-flex`,
     fontFamily: options.headerFontFamily.join(`,`),
     fontWeight: `bold`,
     flexShrink: 0,
-    lineHeight: 1,
+    lineHeight: presets.lineHeights.solid,
+    textDecoration: `none`,
     WebkitFontSmoothing: `antialiased`,
     whiteSpace: `nowrap`,
-    padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
+    padding: `${rhythm(space[2])} ${rhythm(space[3])}`,
     backgroundSize: `30px 30px`,
-    transition: `all ${presets.animation.speedDefault} ${
-      presets.animation.curveDefault
-    }`,
+    transition: `all ${transition.speed.default} ${transition.curve.default}`,
     ":hover, &:focus": {
       backgroundSize: `30px 30px`,
       backgroundColor: colors.gatsby,
       backgroundImage: `linear-gradient(45deg, rgba(0,0,0, 0.1) 25%, transparent 25%, transparent 50%, rgba(0,0,0, 0.1) 50%, rgba(0,0,0, 0.1) 75%, transparent 75%, transparent)`,
-      color: `#fff`,
+      color: colors.white,
       animation: `${stripeAnimation} 2.8s linear infinite`,
     },
     ":focus": {
       outline: 0,
-      boxShadow: `0 0 0 0.2rem ${hex2rgba(colors.lilac, 0.25)}`,
+      boxShadow: `0 0 0 ${rhythm(space[1])} ${hex2rgba(colors.lilac, 0.25)}`,
     },
     ":after": { content: `''`, display: `block` },
     "& svg": { marginLeft: `.2em` },
-    [presets.Md]: {
-      ...scale(1 / 5),
-      padding: `${rhythm(2 / 6)} ${rhythm(3 / 5)}`,
-    },
-    [presets.Xxl]: { padding: `${rhythm(1 / 2)} ${rhythm(1)}` },
   },
   secondary: {
     backgroundColor: `transparent`,
@@ -74,37 +67,21 @@ export const buttonStyles = {
     fontWeight: `normal`,
   },
   large: {
-    // borderRadius: presets.radiusLg,
-    fontSize: scale(1 / 5).fontSize,
-    padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
-    [presets.Md]: {
-      fontSize: scale(2 / 5).fontSize,
-      padding: `${rhythm(2 / 4)} ${rhythm(3 / 5)}`,
-    },
-    [presets.Xxl]: { padding: `${rhythm(1 / 2)} ${rhythm(1)}` },
+    fontSize: presets.scale[4],
+    padding: `${rhythm(space[3])} ${rhythm(space[4])}`,
   },
   small: {
-    fontSize: scale(-1 / 3).fontSize,
-    padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
+    fontSize: presets.scale[1],
+    padding: `${rhythm(space[2])} ${rhythm(space[3])}`,
     [presets.Md]: {
-      fontSize: scale(-1 / 6).fontSize,
-      padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
-    },
-    [presets.Xxl]: {
-      fontSize: scale(-1 / 6).fontSize,
-      padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
+      fontSize: presets.scale[2],
     },
   },
   tiny: {
-    fontSize: scale(-1 / 3).fontSize,
-    padding: `${rhythm(1 / 5)} ${rhythm(1 / 3)}`,
+    fontSize: presets.scale[1],
+    padding: `${rhythm(space[1])} ${rhythm(space[2])}`,
     [presets.Md]: {
-      fontSize: scale(-1 / 4).fontSize,
-      padding: `${rhythm(1 / 5)} ${rhythm(1 / 3)}`,
-    },
-    [presets.Xxl]: {
-      fontSize: scale(-1 / 5).fontSize,
-      padding: `${rhythm(1 / 5)} ${rhythm(1 / 3)}`,
+      fontSize: presets.scale[2],
     },
   },
   ondark: {
@@ -127,7 +104,7 @@ export const svgStyles = {
     "& .svg-fill-gatsby": { fill: colors.gatsby },
     "& .svg-fill-accent": { fill: colors.accent },
     "& .svg-fill-wisteria": { fill: colors.wisteria },
-    "& .svg-fill-brightest": { fill: `#fff` },
+    "& .svg-fill-brightest": { fill: colors.white },
     "& .svg-fill-gradient-accent-white-45deg": {
       fill: `url(#accent-white-45deg)`,
     },
@@ -143,17 +120,56 @@ export const svgStyles = {
 
 // This is an exceptionally bad name
 export const linkStyles = {
-  ...scale(-1 / 6),
-  alignItems: `center`,
+  fontSize: presets.scale[1],
+  lineHeight: presets.lineHeights.solid,
+  padding: `${rhythm(space[3])} 0`,
   "&&": {
     border: 0,
-    boxShadow: `none`,
     color: colors.gray.calm,
     display: `flex`,
     fontWeight: `normal`,
   },
   "&&:hover": {
-    background: `none`,
     color: colors.gatsby,
+  },
+}
+
+export const formInput = {
+  backgroundColor: colors.white,
+  border: `1px solid ${colors.ui.bright}`,
+  borderRadius: radii[1],
+  color: colors.brand,
+  fontFamily: options.headerFontFamily.join(`,`),
+  padding: rhythm(space[3]),
+  verticalAlign: `middle`,
+  transition: `all ${transition.speed.default} ${transition.curve.default}`,
+  "::placeholder": {
+    color: colors.lilac,
+    opacity: 1,
+  },
+}
+
+export const skipLink = {
+  border: 0,
+  color: colors.gatsby,
+  clip: `rect(0 0 0 0)`,
+  height: 1,
+  width: 1,
+  margin: -1,
+  padding: 0,
+  overflow: `hidden`,
+  position: `absolute`,
+  zIndex: 100,
+  fontSize: presets.scale[1],
+  ":focus": {
+    padding: rhythm(presets.space[4]),
+    position: `fixed`,
+    top: rhythm(presets.space[6]),
+    left: rhythm(presets.space[6]),
+    background: colors.white,
+    textDecoration: `none`,
+    width: `auto`,
+    height: `auto`,
+    clip: `auto`,
   },
 }