+++ packages/gatsby-source-contentful/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.38](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.37...gatsby-source-contentful@2.0.38) (2019-03-12)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+## [2.0.37](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.36...gatsby-source-contentful@2.0.37) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+## [2.0.36](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.35...gatsby-source-contentful@2.0.36) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+## [2.0.35](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.34...gatsby-source-contentful@2.0.35) (2019-03-05)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
 ## [2.0.34](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.33...gatsby-source-contentful@2.0.34) (2019-03-04)
 
 **Note:** Version bump only for package gatsby-source-contentful